Release jarink version 0.1.0
The first release of jarink provides the command "brokenlinks",
to scan for broken links.
The output of this command is list of page with its broken links
in JSON format.

This command accept the following options,

`-verbose`::
Print the page that being scanned to standard error.

`-past-result=<path to JSON file>`::
Scan only the pages reported by result from past scan based
on the content in JSON file.
This minimize the time to re-scan the pages once we have fixed the URLs.